Revert "4 add rule"
