+++ backend/src/database/ngs-builder/entities/segment.entity.ts
